requirements.txt
# Core LangChain dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langsmith>=0.1.0

# Ollama integration
ollama>=0.1.7

# GUI
customtkinter>=5.2.0

# Web interface (optional)
streamlit>=1.28.0

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data handling
pandas>=2.0.3

# Build tools (for creating the Mac app)
pyinstaller>=6.0.0
